src/gemini-file-search/requirements.txt
google-genai>=1.30.0
python-dotenv==1.2.1